@@ -1,26 +1,26 @@
-const fs = require("fs");
-const pathUtil = require("path");
+import * as fsPromises from "node:fs/promises";
+import * as pathUtil from "node:path";
 
 /**
  * Recursively read a directory.
  * @param {string} directory
- * @returns {Array<[string, string]>} List of tuples [name, absolutePath].
+ * @returns {Promise<Array<[string, string]>>} List of tuples [name, absolutePath].
  * The return result includes files in subdirectories, but not the subdirectories themselves.
  */
-const recursiveReadDirectory = (directory) => {
+export const recursiveReadDirectory = async (directory) => {
   const result = [];
-  for (const name of fs.readdirSync(directory)) {
+  for (const name of await fsPromises.readdir(directory)) {
     if (name.startsWith(".")) {
       // Ignore .eslintrc.js, .DS_Store, etc.
       continue;
     }
     const absolutePath = pathUtil.join(directory, name);
-    const stat = fs.statSync(absolutePath);
+    const stat = await fsPromises.stat(absolutePath);
     if (stat.isDirectory()) {
       for (const [
         relativeToChildName,
         childAbsolutePath,
-      ] of recursiveReadDirectory(absolutePath)) {
+      ] of await recursiveReadDirectory(absolutePath)) {
         // This always needs to use / on all systems
         result.push([`${name}/${relativeToChildName}`, childAbsolutePath]);
       }
@@ -34,10 +34,11 @@ const recursiveReadDirectory = (directory) => {
 /**
  * Synchronous create a directory and any parents. Does nothing if the folder already exists.
  * @param {string} directory
+ * @returns {Promise<void>}
  */
-const mkdirp = (directory) => {
+export const mkdirp = async (directory) => {
   try {
-    fs.mkdirSync(directory, {
+    await fsPromises.mkdir(directory, {
       recursive: true,
     });
   } catch (e) {
@@ -45,9 +46,4 @@ const mkdirp = (directory) => {
       throw e;
     }
   }
-};
-
-module.exports = {
-  recursiveReadDirectory,
-  mkdirp,
 };
