@@ -115,4 +115,4 @@ const parseMetadata = (extensionCode) => {
   return metadata;
 };
 
-module.exports = parseMetadata;
+export default parseMetadata;
