@@ -1,11 +1,11 @@
-const {
+import {
   transifexApi,
   ORGANIZATION_NAME,
   PROJECT_NAME,
   RUNTIME_RESOURCE,
   METADATA_RESOURCE,
-} = require("./transifex-common");
-const Builder = require("./builder");
+} from "./transifex-common.js";
+import Builder from "./builder.js";
 
 const uploadRuntimeStrings = async (strings) => {
   if (
@@ -52,7 +52,7 @@ const uploadMetadataStrings = async (strings) => {
 const run = async () => {
   console.log("Building...");
   const builder = new Builder();
-  const build = builder.build();
+  const build = await builder.build();
 
   console.log("Generating strings...");
   const l10n = build.generateL10N();
