@@ -10,14 +10,118 @@
     throw new Error("Local Storage must be run unsandboxed");
   }
 
-  const PREFIX = "extensions.turbowarp.org/local-storage:";
-  let namespace = "";
-  const getFullStorageKey = () => `${PREFIX}${namespace}`;
+  const getNamespace = () =>
+    Scratch.vm.runtime.extensionStorage["localstorage"]?.namespace;
+
+  /**
+   * @param {string} newNamespace
+   */
+  const setNamespace = (newNamespace) => {
+    Scratch.vm.runtime.extensionStorage["localstorage"] = {
+      namespace: newNamespace,
+    };
+    Scratch.vm.extensionManager.refreshBlocks("localstorage");
+    readFromStorage();
+  };
+
+  const STORAGE_PREFIX = "extensions.turbowarp.org/local-storage:";
+  const getStorageKey = () => `${STORAGE_PREFIX}${getNamespace()}`;
+
+  /**
+   * Cached in memory for performance.
+   * @type {Record<string, string|number|boolean>}
+   */
+  let namespaceValues = Object.create(null);
+
+  const readFromStorage = () => {
+    namespaceValues = Object.create(null);
+
+    try {
+      // localStorage could throw if unsupported
+      const data = localStorage.getItem(getStorageKey());
+      if (data) {
+        // JSON.parse could throw if data is invalid
+        const parsed = JSON.parse(data);
+        if (parsed && parsed.data) {
+          // Remove invalid values from the JSON
+          for (const [key, value] of Object.entries(parsed.data)) {
+            if (
+              typeof value === "string" ||
+              typeof value === "number" ||
+              typeof value === "boolean"
+            ) {
+              namespaceValues[key] = value;
+            }
+          }
+        }
+      }
+    } catch (error) {
+      console.error("Error reading from local storage", error);
+    }
+  };
+
+  const saveToLocalStorage = () => {
+    try {
+      if (Object.keys(namespaceValues).length > 0) {
+        localStorage.setItem(
+          getStorageKey(),
+          JSON.stringify({
+            // If we find that turbowarp.org is commonly running out of shared space in local storage,
+            // having a timestamp here makes it at least theoretically possible to delete storage based
+            // on last used time.
+            time: Math.round(Date.now() / 1000),
+            data: namespaceValues,
+          })
+        );
+      } else {
+        localStorage.removeItem(getStorageKey());
+      }
+    } catch (error) {
+      console.error("Error saving to local storage", error);
+    }
+  };
+
+  window.addEventListener("storage", (event) => {
+    if (
+      getNamespace() &&
+      event.key === getStorageKey() &&
+      event.storageArea === localStorage
+    ) {
+      readFromStorage();
+      Scratch.vm.runtime.startHats("localstorage_whenChanged");
+    }
+  });
+
+  const generateRandomNamespace = () => {
+    // doesn't need to be cryptographically secure and doesn't need to have excessive length
+    // this has 16^16 = 18446744073709551616 possible namespaces which is plenty
+    const soup = "0123456789abcdef";
+    let id = "";
+    for (let i = 0; i < 16; i++) {
+      id += soup[Math.floor(Math.random() * soup.length)];
+    }
+    return id;
+  };
+
+  const generateRandomNamespaceIfMissing = () => {
+    if (!getNamespace()) {
+      setNamespace(generateRandomNamespace());
+    }
+  };
+
+  Scratch.vm.runtime.on("PROJECT_LOADED", () => {
+    generateRandomNamespaceIfMissing();
+  });
+
+  Scratch.vm.runtime.on("RUNTIME_DISPOSED", () => {
+    generateRandomNamespace();
+  });
+
+  generateRandomNamespaceIfMissing();
 
   let lastNamespaceWarning = 0;
-
   const validNamespace = () => {
-    const valid = !!namespace;
+    const valid = !!getNamespace();
     if (!valid && Date.now() - lastNamespaceWarning > 3000) {
       alert(
         Scratch.translate(
@@ -29,66 +133,6 @@
     return valid;
   };
 
-  const readFromStorage = () => {
-    try {
-      // localStorage could throw if unsupported
-      const data = localStorage.getItem(getFullStorageKey());
-      if (data) {
-        // JSON.parse could throw if data is invalid
-        const parsed = JSON.parse(data);
-        if (parsed && parsed.data) {
-          // Remove invalid values from the JSON
-          const processed = {};
-          for (const [key, value] of Object.entries(parsed.data)) {
-            if (
-              typeof value === "number" ||
-              typeof value === "string" ||
-              typeof value === "boolean"
-            ) {
-              processed[key] = value;
-            }
-          }
-          return processed;
-        }
-      }
-    } catch (error) {
-      console.error("error reading from local storage", error);
-    }
-    return {};
-  };
-
-  const saveToLocalStorage = (data) => {
-    try {
-      if (Object.keys(data).length > 0) {
-        localStorage.setItem(
-          getFullStorageKey(),
-          JSON.stringify({
-            time: Math.round(Date.now() / 1000),
-            data,
-          })
-        );
-      } else {
-        localStorage.removeItem(getFullStorageKey());
-      }
-    } catch (error) {
-      console.error("error saving to locacl storage", error);
-    }
-  };
-
-  window.addEventListener("storage", (event) => {
-    if (
-      namespace &&
-      event.key === getFullStorageKey() &&
-      event.storageArea === localStorage
-    ) {
-      Scratch.vm.runtime.startHats("localstorage_whenChanged");
-    }
-  });
-
-  Scratch.vm.runtime.on("RUNTIME_DISPOSED", () => {
-    namespace = "";
-  });
-
   class LocalStorage {
     getInfo() {
       return {
@@ -97,20 +141,22 @@
         docsURI: "https://extensions.turbowarp.org/local-storage",
         blocks: [
           {
-            opcode: "setProjectId",
-            blockType: Scratch.BlockType.COMMAND,
-            text: Scratch.translate("set storage namespace ID to [ID]"),
-            arguments: {
-              ID: {
-                type: Scratch.ArgumentType.STRING,
-                defaultValue: Scratch.translate("project title"),
-              },
-            },
+            blockType: Scratch.BlockType.LABEL,
+            text: getNamespace()
+              ? Scratch.translate(
+                  {
+                    default: "Namespace: {namespace}",
+                  },
+                  {
+                    namespace: getNamespace(),
+                  }
+                )
+              : Scratch.translate("No namespace set"),
           },
           {
             opcode: "get",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("get key [KEY]"),
+            text: Scratch.translate("get [KEY] from storage"),
             arguments: {
               KEY: {
                 type: Scratch.ArgumentType.STRING,
@@ -121,7 +167,7 @@
           {
             opcode: "set",
             blockType: Scratch.BlockType.COMMAND,
-            text: Scratch.translate("set key [KEY] to [VALUE]"),
+            text: Scratch.translate("set [KEY] to [VALUE] in storage"),
             arguments: {
               KEY: {
                 type: Scratch.ArgumentType.STRING,
@@ -136,7 +182,7 @@
           {
             opcode: "remove",
             blockType: Scratch.BlockType.COMMAND,
-            text: Scratch.translate("delete key [KEY]"),
+            text: Scratch.translate("delete [KEY] from storage"),
             arguments: {
               KEY: {
                 type: Scratch.ArgumentType.STRING,
@@ -147,7 +193,7 @@
           {
             opcode: "removeAll",
             blockType: Scratch.BlockType.COMMAND,
-            text: Scratch.translate("delete all keys"),
+            text: Scratch.translate("delete storage"),
           },
           {
             opcode: "whenChanged",
@@ -155,45 +201,62 @@
             text: Scratch.translate("when another window changes storage"),
             isEdgeActivated: false,
           },
+          "---",
+          {
+            opcode: "setProjectId",
+            blockType: Scratch.BlockType.COMMAND,
+            text: Scratch.translate("set namespace to [ID]"),
+            arguments: {
+              ID: {
+                type: Scratch.ArgumentType.STRING,
+                defaultValue:
+                  getNamespace() || Scratch.translate("project title"),
+              },
+            },
+          },
         ],
       };
     }
+
     setProjectId({ ID }) {
-      namespace = Scratch.Cast.toString(ID);
+      setNamespace(Scratch.Cast.toString(ID));
     }
+
     get({ KEY }) {
       if (!validNamespace()) {
         return "";
       }
-      const storage = readFromStorage();
       KEY = Scratch.Cast.toString(KEY);
-      if (!Object.prototype.hasOwnProperty.call(storage, KEY)) {
+      if (!Object.prototype.hasOwnProperty.call(namespaceValues, KEY)) {
         return "";
       }
-      return storage[KEY];
+      return namespaceValues[KEY];
     }
+
     set({ KEY, VALUE }) {
       if (!validNamespace()) {
         return "";
       }
-      const storage = readFromStorage();
-      storage[Scratch.Cast.toString(KEY)] = VALUE;
-      saveToLocalStorage(storage);
+      namespaceValues[Scratch.Cast.toString(KEY)] = VALUE;
+      saveToLocalStorage();
     }
+
     remove({ KEY }) {
       if (!validNamespace()) {
         return "";
       }
-      const storage = readFromStorage();
-      delete storage[Scratch.Cast.toString(KEY)];
-      saveToLocalStorage(storage);
+      delete namespaceValues[Scratch.Cast.toString(KEY)];
+      saveToLocalStorage();
     }
+
     removeAll() {
       if (!validNamespace()) {
         return "";
       }
-      saveToLocalStorage({});
+      namespaceValues = Object.create(null);
+      saveToLocalStorage();
     }
   }
+
   Scratch.extensions.register(new LocalStorage());
 })(Scratch);
